R/beatmaps: update the params insertion rule
Beatmap id is unique enough for query a beatmap. For user who provide
beatmap id, they may not want to see tons of result because of the set
id is provided. So here I add a new condition that the cli only insert
set id into query params when their is no beatmap id provided.

Signed-off-by: Avimitin <avimitin@gmail.com>